extract each lint into its own module
